third_party/blink/web_tests/platform/generic/http/tests/security/contentSecurityPolicy/source-list-parsing-nonascii-expected.txt

CONSOLE ERROR: The value for the Content-Security-Policy directive 'script-src' contains one or more invalid characters. In a source expression, non-whitespace characters outside ASCII 0x21-0x7E must be Punycode-encoded, as described in RFC 3492 (https://tools.ietf.org/html/rfc3492), if part of the hostname and percent-encoded, as described in RFC 3986, section 2.1 (http://tools.ietf.org/html/rfc3986#section-2.1), if part of the path.
Sources containing non-ascii characters should be ignored, and should generate warnings.

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
